spreadout funcs and configs
