+++ STM32_workspace_9.1/Project 0 Code.zip_expanded/Project0_Code/src/main.c
@@ -1,35 +1,7 @@
-
-/* Standard includes. */
-#include <stdint.h>
-#include <stdio.h>
-#include "stm32f4_discovery.h"
-/* Kernel includes. */
-#include "stm32f4xx.h"
-#include "../FreeRTOS_Source/include/FreeRTOS.h"
-#include "../FreeRTOS_Source/include/queue.h"
-#include "../FreeRTOS_Source/include/semphr.h"
-#include "../FreeRTOS_Source/include/task.h"
-#include "../FreeRTOS_Source/include/timers.h"
-
-
+#include "dd_tasks.h"
+#include "define_configs.h"
 
 /*-----------------------------------------------------------*/
-
-struct dd_task {
-	TaskHandle_t t_handle;
-	uint32_t type;
-	uint32_t task_id;
-	uint32_t release_time;
-	uint32_t absolute_deadline;
-	uint32_t execution_time;
-	uint32_t completion_time;
-	uint32_t message;
-};
-
-struct dd_task_list {
-	struct dd_task task;
-	struct dd_task_list* next_task;
-};
 
 /*
  * TODO: Implement this function for any hardware specific clock configuration
@@ -40,21 +12,8 @@ static void DD_Task_Generator( void *pvParameters);
 static void Create_UD_Task(void *pvParameters);
 static void Deadline_Driven_Scheduler(void *pvParameters);
 static void DD_Task_Monitor( void *pvParameters);
-typedef enum {PERIODIC, APERIODIC} task_type;
 
-void addToList(struct dd_task_list*, struct dd_task, int);
-void dd_remove_overdue(struct dd_task_list*, struct dd_task_list*);
-struct dd_task dd_delete(struct dd_task_list*, uint32_t);
-int get_active_dd_task_list(struct dd_task_list*);
-int get_complete_dd_task_list(struct dd_task_list*);
-int get_overdue_dd_task_list(struct dd_task_list*);
-void create_dd_task(uint32_t , uint32_t , uint32_t, uint32_t );
 
-xQueueHandle active_task_queue = 0;
-xQueueHandle complete_task_queue = 0;
-xQueueHandle overdue_task_queue = 0;
-xQueueHandle Complete_Queue = 0;
-xQueueHandle Create_Queue = 0;
 
 /*-----------------------------------------------------------*/
 
@@ -262,163 +221,7 @@ void DD_Task_Generator( void *pvParameters ){
 	}
 }
 
-//void dd_start(struct dd_task_list* head){
-//	head->next_task->task.release_time = (uint32_t)xTaskGetTickCount();
-//	printf("Released task: %d at: %d s\n", (int)head->next_task->task.task_id, (int)head->next_task->task.release_time);
-//	xTaskCreate(Create_UD_Task, "UD Task", configMINIMAL_STACK_SIZE, &head->next_task->task, 3, &(head->next_task->task.t_handle));
-//}
 
-void Create_UD_Task(void *pvParameters){
-//	printf("Creating UD task\n");
-	struct dd_task* curr_task = (struct dd_task*) pvParameters;
-//	Get tick count
-	int start = (int)xTaskGetTickCount();
-	int end = start + (int) curr_task->execution_time;
-
-//	Keep running before task ends
-	while ((int)xTaskGetTickCount() < end);
-	uint32_t task_id = curr_task->task_id;
-	curr_task->completion_time = (uint32_t) xTaskGetTickCount();
-
-//	Complete task and send to queue
-	if(!xQueueSend(Complete_Queue, &task_id, 0)) {
-			printf("Failed to send completed task to queue \n");
-	}
-	vTaskSuspend(NULL);
-}
-
-
-
-void create_dd_task(uint32_t type, uint32_t task_id, uint32_t execution_time, uint32_t absolute_deadline) {
-
-	struct dd_task new_dd_task;
-	new_dd_task.type = type;
-	new_dd_task.task_id = task_id;
-	new_dd_task.execution_time = execution_time;
-	new_dd_task.absolute_deadline = absolute_deadline;
-	new_dd_task.release_time = -1;
-	new_dd_task.completion_time = -1; //use l8er
-
-	if( !xQueueSend( Create_Queue, &new_dd_task, 0)) {
-		printf("Failed to create a task\n");
-	}
-
-}
-
-
-struct dd_task dd_delete(struct dd_task_list* head, uint32_t task_id){
-//	printf("Deleting a deadly driven task\n");
-	struct dd_task_list *current = head;
-	struct dd_task_list *previous = head;
-
-	while(current->task.task_id != task_id) {
-		previous = current;
-		current = current->next_task;
-	}
-	if(current->next_task == NULL) {
-		previous->next_task = NULL;
-	}
-	else {
-		previous->next_task = current->next_task;
-	}
-	vTaskDelete(current->task.t_handle);
-	return current->task;
-}
-
-// TASK LISTS
-/*-----------------------------------------------------------*/
-int get_active_dd_task_list(struct dd_task_list* head) {
-
-	struct dd_task_list *current = head;
-	int i = 0;
-
-	while(current->next_task != NULL) {
-		current = current->next_task;
-
-		i++;
-	}
-
-    return i;
-}
-
-int get_complete_dd_task_list(struct dd_task_list* head) {
-	struct dd_task_list *current = head;
-	int i = 0;
-
-	while(current->next_task != NULL) {
-		current = current->next_task;
-
-		i++;
-	}
-	return i;
-
-
-}
-
-int get_overdue_dd_task_list(struct dd_task_list* head) {
-	struct dd_task_list *current = head;
-	int i = 0;
-
-	while(current->next_task != NULL) {
-		current = current->next_task;
-
-		i++;
-	}
-	return i;
-
-}
-
-void dd_remove_overdue(struct dd_task_list* activeHead, struct dd_task_list* overdueHead){
-	struct dd_task_list *current = activeHead;
-	struct dd_task_list *previous = activeHead;
-
-	while (current != NULL) {
-		while (current != NULL && current->task.absolute_deadline > (uint32_t)xTaskGetTickCount()){
-			previous = current;
-			current = current->next_task;
-		}
-		if(current->task.release_time != -1 && current->task.completion_time == -1){
-			previous = current;
-			current = current->next_task;
-			continue;
-		}
-		if (current == NULL) {
-			return;
-		}
-
-		previous->next_task = current->next_task;
-		addToList(overdueHead, current->task, 0);
-		printf("Overdue Task: %d at: %d\n", (int)current->task.task_id, (int) xTaskGetTickCount());
-		current = previous->next_task;
-	}
-}
-
-void addToList(struct dd_task_list *head, struct dd_task next_task, int f_sort) {
-	struct dd_task_list *new_node = (struct dd_task_list*)malloc(sizeof(struct dd_task_list));
-	struct dd_task_list *current = head;
-
-	new_node->task = next_task;
-	new_node->next_task = NULL;
-
-	if(f_sort == 1) {
-		while(current->next_task != NULL) {
-			if (new_node->task.absolute_deadline < current->next_task->task.absolute_deadline) {
-				new_node->next_task = current->next_task;
-				current->next_task = new_node;
-				return;
-			} else {
-				current = current->next_task;
-			}
-		}
-	}
-	else {
-		while(current->next_task != NULL) {
-			current = current->next_task;
-		}
-	}
-	current->next_task = new_node;
-
-}
 
 
 /*-----------------------------------------------------------*/
